- Refactors file management helpers out into own module
